Premium Analytics: move the analytics URL helper into the Jetpack plugin
The helper was in js-packages/script-data, which meant adding date-fns to
a previously zero-dependency package that 23 projects depend on and that
is autopublished to npm. It only ever had two consumers, and one of them
did not need it at all.

My Jetpack drops it entirely. Its Stats card already receives the
destination as the product's `manage_url`, which `Stats::get_manage_url()`
resolves server-side, so the JS was rebuilding a URL PHP had already
sent. The one thing the URL cannot carry is whether to append
`force_refresh=1` — that is now a `premiumAnalyticsEnabled` entry in
`get_my_jetpack_flags()`, alongside the existing showFullJetpackStatsCard
and isStatsModuleActive flags.

That leaves the Jetpack plugin as the only consumer, so the helper moves
to `_inc/shared/analytics-url.ts`. Its three call sites live in different
trees built by two webpack configs, and the extensions build has no
bare-specifier root, so each imports it by relative path. `_inc/shared/`
is new: `_inc/client`, `modules/` and `extensions/` had no shared JS
between them before this.

script-data goes back to exactly trunk apart from the `analytics` key's
type — the shape of the global is its job, building URLs from it is not.
The jest bootstrap it needed goes away with the test, which now runs
under the plugin's existing client config.
