+++ projects/js-packages/script-data/composer.json
@@ -4,10 +4,7 @@
 	"type": "library",
 	"license": "GPL-2.0-or-later",
 	"require": {},
-	"scripts": {
-		"test-js": "pnpm run test",
-		"test-js-coverage": "pnpm run test-coverage"
-	},
+	"scripts": {},
 	"repositories": [
 		{
 			"type": "path",
