--- projects/js-packages/script-data/jest.config.cjs
@@ -1,5 +1,0 @@
-const baseConfig = require( 'jetpack-js-tools/jest/config.base.js' );
-
-module.exports = {
-	...baseConfig,
-};
