+++ projects/js-packages/script-data/package.json
@@ -17,20 +17,9 @@
 		".": "./src/index.ts"
 	},
 	"scripts": {
-		"test": "NODE_OPTIONS=--experimental-vm-modules jest",
-		"test-coverage": "pnpm run test --coverage",
 		"typecheck": "tsgo --noEmit"
 	},
-	"dependencies": {
-		"@date-fns/tz": "1.4.1",
-		"date-fns": "4.1.0"
-	},
 	"devDependencies": {
-		"@babel/core": "7.29.7",
-		"@jest/globals": "30.4.1",
-		"@types/jest": "30.0.0",
-		"@typescript/native-preview": "7.0.0-dev.20260707.2",
-		"jest": "30.4.2",
-		"jetpack-js-tools": "workspace:*"
+		"@typescript/native-preview": "7.0.0-dev.20260707.2"
 	}
 }
