+++ projects/js-packages/script-data/src/index.ts
@@ -1,3 +1,2 @@
-export * from './analytics-url.ts';
 export * from './types.ts';
 export * from './utils.ts';
