+++ projects/js-packages/script-data/src/types.ts
@@ -54,8 +54,11 @@ export interface UserData {
 /**
  * Analytics dashboard data, published by the Premium Analytics package only on
  * sites where it *is* the analytics UI. Its absence means the site still runs
- * the legacy Stats dashboard, so `getAnalyticsUrl()` treats presence as the
- * branch signal rather than taking a flag from every caller.
+ * the legacy Stats dashboard, so consumers treat presence as the branch signal
+ * rather than taking a flag from every caller.
+ *
+ * Only the shape of the global lives here, which is this package's job. Building
+ * URLs from it is the Jetpack plugin's — see `_inc/shared/analytics-url.ts`.
  */
 export interface AnalyticsScriptData {
 	/**
@@ -76,51 +79,11 @@ export interface AnalyticsScriptData {
 
 	/**
 	 * The site's timezone — an IANA name (`America/New_York`) when one is set,
-	 * otherwise a fixed UTC offset (`+05:30`). Used to encode date ranges the
-	 * way the dashboard writes them to its own URL.
+	 * otherwise a fixed UTC offset (`+05:30`). Used to encode date ranges the way
+	 * the dashboard writes them to its own URL.
 	 */
 	timezone: string;
 }
-
-/**
- * A range of whole calendar days in the site's timezone, as `YYYY-MM-DD`.
- * Callers pass plain calendar dates; `getAnalyticsUrl()` owns the conversion to
- * whatever encoding the destination expects.
- */
-export interface AnalyticsDateRange {
-	from: string;
-	to: string;
-}
-
-/**
- * A section of the analytics dashboard, in caller-facing vocabulary.
- */
-export type AnalyticsDashboardSection = 'traffic' | 'insights' | 'subscribers' | 'store';
-
-/**
- * A tab of the single-post analytics view, in caller-facing vocabulary.
- */
-export type AnalyticsPostSection = 'traffic' | 'email-opens' | 'email-clicks';
-
-/**
- * An analytics destination, described in terms of what the user should see
- * rather than how the URL is spelled.
- *
- * A discriminated union rather than a property bag, so each view only accepts
- * the modifiers that apply to it. `range` is accepted on both because the
- * dashboard and the post detail page read the same date-window params.
- *
- * Only the two views Jetpack actually links to are modelled. The dashboard also
- * serves `/reports/$report` and `/video/$videoId`; add them here when something
- * needs to link there, rather than ahead of a caller.
- *
- * Deliberately absent: the site and blog identifiers (resolved from script
- * data, so callers need not thread them through props) and product checkout,
- * which shares no path grammar with the analytics views.
- */
-export type AnalyticsView =
-	| { view: 'dashboard'; section?: AnalyticsDashboardSection; range?: AnalyticsDateRange }
-	| { view: 'post'; id: number; section?: AnalyticsPostSection; range?: AnalyticsDateRange };
 
 export interface JetpackScriptData {
 	site: SiteData;
