+++ projects/packages/my-jetpack/src/class-initializer.php
@@ -528,6 +528,11 @@ class Initializer {
 		$flags = array(
 			'videoPressStats'          => Jetpack_Constants::is_true( 'JETPACK_MY_JETPACK_VIDEOPRESS_STATS_ENABLED' ),
 			'showFullJetpackStatsCard' => class_exists( 'Jetpack' ),
+			// The Stats card's destination already arrives as the product's
+			// `manage_url`. This says which destination it is, which is only needed
+			// because the legacy Stats page caches its report and wants a
+			// `force_refresh` hint the dashboard has no use for.
+			'premiumAnalyticsEnabled'  => Products\Stats::is_premium_analytics_enabled(),
 		);
 
 		return $flags;
