+++ projects/plugins/jetpack/package.json
@@ -60,6 +60,7 @@
 		"@automattic/popup-monitor": "1.0.2",
 		"@automattic/social-previews": "workspace:*",
 		"@automattic/viewport": "1.0.0",
+		"@date-fns/tz": "1.4.1",
 		"@react-spring/web": "10.1.1",
 		"@visx/curve": "4.0.0",
 		"@visx/legend": "^4.0.0",
@@ -95,6 +96,7 @@
 		"cookie": "1.0.2",
 		"copy-webpack-plugin": "14.0.0",
 		"crypto-js": "4.2.0",
+		"date-fns": "4.1.0",
 		"debounce": "2.2.0",
 		"debug": "4.4.3",
 		"deepmerge": "4.3.1",
